pyproject.toml
[tool.poetry]
name = "django-ecommerce"
version = "0.1.0"
description = ""
authors = ["mostafaei2002 <sinamostafaei2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-ckeditor = "^6.7.0"
django-htmx = "^1.17.2"
faker = "^22.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
